choi - resume 수정

+++ SQL/PROCEDURE/resume_remove.sql
@@ -2,12 +2,18 @@
 -- ✔️ 회사 계정 → 공고 삭제
 delimiter //
 create procedure RESUME_003(
+    in memberIdInput bigint, 
     in idInput bigint
 )
 begin
-    delete from resume
-    where id = idInput;
+    if exists(select 1 from member where id = memberIdInput and member_type = 'user' and state = 'online') then
+        delete from resume
+        where id = idInput;
 
-    select '이력서가 삭제되었습니다.' as message;
+        select '이력서가 삭제되었습니다.' as message;
+    else
+        -- 개인 계정이 아닌 경우 오류 발생
+        signal sqlstate '45000' set message_text = '개인 계정 및 online 상태만 삭제 가능!';
+    end if;
 end //
 delimiter ;
